refactor: store 패키지 위치 리팩터링
