+++ domain/domain-pos/src/main/java/domain/pos/store/implement/SaleReader.java
@@ -6,7 +6,7 @@ import org.springframework.data.domain.Slice;
 import org.springframework.stereotype.Component;
 
 import domain.pos.store.entity.Sale;
-import domain.pos.store.repository.SaleRepository;
+import domain.pos.store.port.required.SaleRepository;
 import lombok.RequiredArgsConstructor;
 
 @Component
