+++ domain/domain-pos/src/main/java/domain/pos/store/implement/StoreReader.java
@@ -10,7 +10,7 @@ import com.vo.UserPassport;
 
 import domain.pos.store.entity.Store;
 import domain.pos.store.entity.dto.StoreHeadDto;
-import domain.pos.store.repository.StoreRepository;
+import domain.pos.store.port.required.StoreRepository;
 import lombok.RequiredArgsConstructor;
 
 @Component
