+++ domain/domain-pos/src/main/java/domain/pos/store/implement/StoreWriter.java
@@ -6,7 +6,7 @@ import com.vo.UserPassport;
 
 import domain.pos.store.entity.Store;
 import domain.pos.store.entity.StoreInfo;
-import domain.pos.store.repository.StoreRepository;
+import domain.pos.store.port.required.StoreRepository;
 import lombok.RequiredArgsConstructor;
 
 @Component
