+++ domain/domain-pos/src/main/java/domain/pos/store/port/required/PresignedUrlGenerator.java
@@ -1,4 +1,4 @@
-package domain.pos.store.repository;
+package domain.pos.store.port.required;
 
 public interface PresignedUrlGenerator {
 	String generatePresignedUrl(String s3Path);
