+++ infra/rdb/pos/src/main/java/com/pos/sale/repository/SaleRepositoryImpl.java
@@ -18,7 +18,7 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 
 import domain.pos.store.entity.Sale;
 import domain.pos.store.entity.Store;
-import domain.pos.store.repository.SaleRepository;
+import domain.pos.store.port.required.SaleRepository;
 import lombok.RequiredArgsConstructor;
 
 @Repository
