+++ infra/rdb/pos/src/main/java/com/pos/store/repository/StoreRepositoryImpl.java
@@ -21,7 +21,7 @@ import com.vo.UserPassport;
 import domain.pos.store.entity.Store;
 import domain.pos.store.entity.StoreInfo;
 import domain.pos.store.entity.dto.StoreHeadDto;
-import domain.pos.store.repository.StoreRepository;
+import domain.pos.store.port.required.StoreRepository;
 import jakarta.persistence.LockModeType;
 import lombok.RequiredArgsConstructor;
