+++ infra/rdb/pos/src/test/java/com/pos/store/repository/StoreRepositoryImplTest.java
@@ -28,7 +28,7 @@ import com.pos.table.entity.TableEntity;
 import domain.pos.store.entity.Store;
 import domain.pos.store.entity.StoreInfo;
 import domain.pos.store.entity.dto.StoreHeadDto;
-import domain.pos.store.repository.StoreRepository;
+import domain.pos.store.port.required.StoreRepository;
 
 class StoreRepositoryImplTest extends RepositoryTest {
