+++ settings.gradle
@@ -20,7 +20,6 @@ include(":common:domain-entity")
 // domain
 include(":domain:domain-pos")
 include(":domain:domain-event:store-event")
-include(":domain:domain-store")
 
 // infra
 include(":infra:rdb:pos")
